Fix IAR build issues

+++ libraries/c_sdk/standard/mqtt/test/unit/iot_tests_mqtt_api.c
@@ -544,9 +544,9 @@ static int32_t transportSend( const NetworkContext_t * pNetworkContext,
 /**
  * @brief A transport send function that delays.
  */
-static size_t transportSendDelay( const NetworkContext_t * pSendContext,
-                                  const void * pMessage,
-                                  size_t messageLength )
+static int32_t transportSendDelay( const NetworkContext_t * pSendContext,
+                                   const void * pMessage,
+                                   size_t messageLength )
 {
     IotSemaphore_t * pWaitSem = ( IotSemaphore_t * ) pSendContext;
